+++ src/romeo/base-page.js
@@ -5,7 +5,7 @@ const DEFAULT_OPTIONS = {
   index: 1,
   isCurrent: true,
   queue: null,
-  seed: null,
+  guard: null,
   iota: null
 };
 
@@ -24,27 +24,30 @@ class BasePage extends Base {
     this.addresses = {};
   }
 
-  async init() {
-    return await this.sync();
+  async init(force = false, priority) {
+    return await this.sync(force, priority);
   }
 
   async sync(force = false, priority) {
-    await this.syncAddresses(priority, force);
+    const { index } = this.opts;
+    if (!priority) {
+      priority = index + 1;
+    }
+    await this.syncAddresses(priority, !force);
     if (!Object.keys(this.addresses).length) {
       await this.getNewAddress();
-      await this.syncAddresses(priority, true);
     }
     return this;
   }
 
   asJson() {
-    const { index, isCurrent, seed } = this.opts;
+    const { index, isCurrent, guard } = this.opts;
     return {
       index,
       isCurrent,
-      seed,
-      addresses: this.addresses,
-      jobs: this.getJobs()
+      seed: guard.getPageSeed(index),
+      addresses: Object.assign({}, this.addresses),
+      jobs: this.getJobs().map(j => Object.assign({}, j))
     };
   }
 
@@ -76,30 +79,31 @@ class BasePage extends Base {
   }
 
   getNewAddress(total = 1, callback = null) {
-    const { iota, seed, queue, index, isCurrent } = this.opts;
+    const { iota, queue, index, isCurrent } = this.opts;
 
     return new Promise((resolve, reject) => {
       const addressPromise = () =>
         new Promise((resolve, reject) => {
-          iota.api.getNewAddress(
-            seed,
-            {
-              index: Object.keys(this.addresses).length,
-              total
-            },
+          iota.api.ext.getNewAddress(
+            index,
+            Object.keys(this.addresses).length,
+            total,
             async (err, addresses) => {
               if (err) {
                 reject(err);
               }
               addresses = Array.isArray(addresses) ? addresses : [addresses];
-              this.applyAddresses(addresses);
+              await this.applyAddresses(addresses);
               await this.restoreAddresses(
                 addresses,
                 `Attaching new addresses`,
                 'Could not attach new addresses'
               );
-              await this.syncAddresses(100);
-              callback && (await callback(addresses));
+              await iota.api.ext.setAddresses(
+                index,
+                Object.keys(this.addresses)
+              );
+              callback && callback(addresses);
               resolve(addresses);
             }
           );
@@ -110,16 +114,20 @@ class BasePage extends Base {
         type: 'NEW_ADDRESS',
         description: `Adding new addresses`
       });
-      job.on('finish', resolve);
+      job.on('finish', result => {
+        this.onChange();
+        resolve(result);
+      });
       job.on('failed', err => {
         this.log('Could not add addresses', err);
         reject(err);
       });
+      this.onChange();
     });
   }
 
-  syncAddresses(priority, cachedOnly) {
-    const { iota, seed, queue, index, isCurrent } = this.opts;
+  syncAddresses(priority, cachedOnly, total) {
+    const { iota, queue, index, isCurrent } = this.opts;
 
     return new Promise((resolve, reject) => {
       let cached = [];
@@ -127,7 +135,7 @@ class BasePage extends Base {
       const addressPromise = () =>
         new Promise((resolve, reject) => {
           iota.api.ext.getAddresses(
-            seed,
+            index,
             (err, addresses) => {
               if (!err) {
                 cached = addresses;
@@ -148,7 +156,8 @@ class BasePage extends Base {
                 this.applyAddresses(addresses);
               resolve(this);
             },
-            cachedOnly
+            cachedOnly,
+            total
           );
         });
 
@@ -162,25 +171,40 @@ class BasePage extends Base {
           cachedOnly
         }
       );
-      job.on('finish', resolve);
+      job.on('finish', result => {
+        this.onChange();
+        resolve(result);
+      });
       job.on('failed', err => {
         this.log('Could not sync page addresses', err);
         reject(err);
       });
+      this.onChange();
     });
   }
 
-  sendTransfers(transfers, inputs, message, messageFail, priority) {
-    const { iota, seed, queue, index, isCurrent } = this.opts;
+  sendTransfers(
+    transfers,
+    inputs,
+    message,
+    messageFail,
+    priority,
+    preventRetries,
+    options
+  ) {
+    const { iota, queue, index, isCurrent } = this.opts;
 
     const sendPromise = () =>
       new Promise((resolve, reject) => {
-        iota.api.sendTransfer(
-          seed,
+        iota.api.ext.sendTransfer(
+          index,
           IOTA_DEPTH,
           IOTA_MWM,
           transfers,
-          { inputs },
+          Object.assign({}, options, {
+            inputs,
+            addressIndex: Object.keys(this.addresses).length
+          }),
           (err, result) => {
             if (err) {
               return reject(err);
@@ -197,7 +221,8 @@ class BasePage extends Base {
         {
           page: index,
           type: 'SEND_TRANSFER',
-          description: message || `Sending transfers`
+          description: message || `Sending transfers`,
+          preventRetries
         }
       );
       job.on('finish', resolve);
@@ -205,22 +230,18 @@ class BasePage extends Base {
         this.log(messageFail || 'Could not send transfer to the tangle', err);
         reject(err);
       });
+      this.onChange();
     });
   }
 
   restoreAddresses(addresses, message, messageFail) {
     message = message || `Restoring addresses`;
     messageFail = messageFail || 'Could not restore the addresses';
-    return Promise.all(
-      addresses.map(
-        async address =>
-          await this.sendTransfers(
-            [{ address, value: 0 }],
-            null,
-            message,
-            messageFail
-          )
-      )
+    return this.sendTransfers(
+      addresses.map(address => ({ address, value: 0 })),
+      null,
+      message,
+      messageFail
     ).then(res => {
       this.onChange();
       return res;