+++ src/romeo/romeo.js
@@ -1,15 +1,13 @@
 const { Base } = require('./base');
-const crypto = require('../crypto');
 const createQueue = require('../queue');
-const createAPI = require('../iota');
 const { Database } = require('../db');
 const { Pages } = require('./pages');
 
 const DEFAULT_OPTIONS = {
-  username: null,
-  password: null,
   syncInterval: 60000,
-  dbPath: 'romeo'
+  dbPath: 'romeo',
+  guard: null,
+  account: 0
 };
 
 class Romeo extends Base {
@@ -23,16 +21,21 @@ class Romeo extends Base {
       options
     );
     super(opts);
+    if (!this.opts.guard) throw new Error('No guard provided!');
+    this.guard = this.opts.guard;
     this.ready = false;
     this.isOnline = 1;
     this.checkingOnline = false;
+    this.addingPage = false;
     this.opts = opts;
-    this.keys = crypto.keys.getKeys(opts.username, opts.password);
-    this.db = new Database({ path: opts.dbPath, password: this.keys.password });
-    this.iota = createAPI({ database: this.db });
+    this.db = new Database({
+      path: `${opts.dbPath}-account-${this.guard.opts.account}`,
+      password: this.guard.getSymmetricKey()
+    });
+    this.iota = this.guard.setupIOTA({ database: this.db });
     this.queue = createQueue();
     this.pages = new Pages({
-      keys: this.keys,
+      guard: this.guard,
       queue: this.queue,
       iota: this.iota,
       db: this.db,
@@ -49,7 +52,7 @@ class Romeo extends Base {
     if (restoreString) {
       await this.db.restore(restoreString, true);
     }
-    await this.pages.init();
+    await this.pages.init(false, 10000);
     this.updater = setInterval(
       () => this.pages.syncCurrentPage(),
       this.opts.syncInterval
@@ -75,23 +78,16 @@ class Romeo extends Base {
   }
 
   asJson() {
-    const {
-      queue: { jobs },
-      keys,
-      pages,
-      isOnline,
-      checkingOnline,
-      ready
-    } = this;
+    const { queue: { jobs }, pages, isOnline, checkingOnline, ready } = this;
     return {
-      keys,
-      jobs: Object.values(jobs),
-      genericJobs: pages.getJobs(),
+      jobs: Object.values(jobs).map(j => Object.assign({}, j)),
+      genericJobs: pages.getJobs().map(j => Object.assign({}, j)),
       pages: pages.asJson(),
       isOnline,
       checkingOnline,
       provider: this.iota.api.ext.provider,
-      ready
+      ready,
+      addingPage: this.addingPage
     };
   }
 
@@ -114,30 +110,60 @@ class Romeo extends Base {
     return await this.db.backup(true);
   }
 
-  async newPage(opts = {}) {
-    const { sourcePage, includeReuse = false } = opts;
-    const currentPage = sourcePage || this.pages.getCurrent();
+  async newPage(opts = {}, onCreate) {
+    this.addingPage = true;
+    try {
+      const { preventRetries, sourcePage, includeReuse = false } = opts;
+      const currentPage = sourcePage || this.pages.getCurrent();
 
-    const newPage = this.pages.getByAddress((await this.pages.getNewPage())[0])
-      .page;
+      const newPage = this.pages.getByAddress(
+        (await this.pages.getNewPage())[0]
+      ).page;
 
-    if (!currentPage.isSynced()) {
-      await currentPage.sync();
+      if (!currentPage.isSynced()) {
+        await currentPage.sync();
+      }
+      const address = newPage.getCurrentAddress().address;
+      const inputs = currentPage.getInputs(includeReuse);
+      const tag = '99ROMEO9NEW9PAGE9TRANSFER99';
+
+      if (this.guard.opts.sequentialTransfers) {
+        for (let input of inputs) {
+          const value = input.balance;
+          await currentPage.sendTransfers(
+            [{ address, value, tag }],
+            [input],
+            'Moving funds to the new page sequentially.',
+            'Failed moving all or some funds!',
+            null,
+            preventRetries
+          );
+        }
+      } else {
+        const value = inputs.reduce((t, i) => t + i.balance, 0);
+        if (value > 0) {
+          await currentPage.sendTransfers(
+            [{ address, value, tag }],
+            inputs,
+            'Moving funds to the new page',
+            'Failed moving funds!',
+            null,
+            preventRetries
+          );
+        }
+      }
+
+      newPage.syncTransactions();
+      onCreate && onCreate(newPage);
+      this.addingPage = false;
+
+      currentPage.syncTransactions();
+      this.onChange();
+      return newPage;
+    } catch (e) {
+      this.addingPage = false;
+      throw e;
     }
-    const address = newPage.getCurrentAddress().address;
-    const inputs = currentPage.getInputs(includeReuse);
-    const value = inputs.reduce((t, i) => t + i.balance, 0);
-    if (value > 0) {
-      await currentPage.sendTransfers(
-        [{ address, value }],
-        inputs,
-        'Moving funds from the current page to the new one',
-        'Failed moving funds from the current page to the new one'
-      );
-      await newPage.syncTransactions();
-    }
-    this.onChange();
-    return newPage;
   }
 
   onChange() {