+++ CHANGELOG.md
@@ -7,6 +7,34 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.5.0] - 2018-10-03
+
+### Added
+
+* Multi-account support for Ledger
+
+## [0.4.5] - 2018-10-03
+
+### Changed
+
+* Multiple performance improvements and fixes
+
+## [0.2.1] - 2018-09-30
+
+### Changed
+
+* Fixes
+* Ledger support
+* Performance improvements
+
+## [0.2.0] - 2018-09-30
+
+### Changed
+
+* Fixes
+* Ledger support
+* Performance improvements
+
 ## [0.1.7] - 2018-09-03
 
 ### Changed